# resources/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/pagmspaypaygos/v1/payments": {
      "post": {
        "operationId": "PaymentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad request"
          }
        },
        "tags": [
          "payments"
        ]
      },
      "get": {
        "operationId": "PaymentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Found all payments"
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
    "/pagmspaypaygos/v1/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      },
      "delete": {
        "operationId": "PaymentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
    "/pagmspaypaygos/v1/services": {
      "post": {
        "operationId": "ServicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Services"
        ]
      },
      "get": {
        "operationId": "ServicesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Services"
        ]
      }
    },
    "/pagmspaypaygos/v1/services/{id}": {
      "get": {
        "operationId": "ServicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Services"
        ]
      },
      "patch": {
        "operationId": "ServicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Services"
        ]
      },
      "delete": {
        "operationId": "ServicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Services"
        ]
      }
    }
  },
  "info": {
    "title": "MS Skeleton",
    "description": "MS Skeleton endpoints",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "https://bancamovil.synopsis.cloud",
      "description": "Production server"
    },
    {
      "url": "http://localhost:${{ values.serverPort }}",
      "description": "Local development"
    }
  ],
  "components": {
    "schemas": {
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User id",
            "nullable": false,
            "minLength": 1
          },
          "serviceId": {
            "type": "string",
            "description": "Service id",
            "nullable": false,
            "minLength": 1
          },
          "amount": {
            "type": "number"
          },
          "paymentMethod": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "serviceId",
          "amount",
          "paymentMethod"
        ]
      },
      "CreateServiceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}